Move src/es/*.js to src/*.mjs
