+++ src/lzma-c.mjs
@@ -1,7 +1,7 @@
 import {
     action_compress,
     compress
-} from './lzma_worker.js';
+} from './lzma_worker';
 
 export {
     compress
