+++ src/lzma-d.mjs
@@ -1,7 +1,7 @@
 import {
     action_decompress,
     decompress
-} from './lzma_worker.js';
+} from './lzma_worker';
 
 export {
     decompress
