+++ src/lzma_worker.mjs
